+++ SpotifyUserLogin/src/main/java/api/SpotifyInteractor.java
@@ -1,4 +1,4 @@
-package apiNew;
+package api;
 
 import org.apache.hc.core5.http.ParseException;
 import se.michaelthelin.spotify.SpotifyApi;
