+++ SpotifyUserLogin/src/main/java/api/Utility.java
@@ -1,4 +1,4 @@
-package apiNew;
+package api;
 
 public class Utility {
     public static String spliceString(char first, char last) {
